Bug in compile rgbmatrix lib,  reworked TCP Socket Code
Forgot a define in the makery for the library

Reworked TCP code so that it will open a connection one time
for all demos and then just send frames.  Server will only stop
listening for data on the connection when it is told to over
the control channel (UDP)

Tested locally with 127.0.0.1:300 server address.  Needs some testing
on a real lan.